Functionalized creating of phylogenetic guide tree.
Changes:
Decoupled functions from global variable and had the variables passed as function parameters.
Moved the code that converts the distance matrix as array of arrays to a matrix in Stefanie's findDistances() function.

Non functional changes:
Changed indentations.

TODO:
Change the function names and variable names to be more intuitive eg. function q on line 23 of the last part.
Change function name findDistances to getDistanceMatrix.
Add more comments.

Errors:
Adding the function create_phylogenetic_guide_tree is giving an error "Cannot `convert` an object of type Main.TreeMod.Tree to
an object of type Main.TreeMod.Tree" which is stupid because both the objects are of the same type (to be inferred as Julia is awesome!!!!).