Functionalized creating of phylogenetic guide tree.
Changes:
Decoupled functions from global variable and had the variables passed as function parameters.
Moved the code that converts the distance matrix as array of arrays to a matrix in Stefanie's findDistances() function.

Non functional changes:
Changed indentations.

TODO:
Change the function names and variable names to be more intuitive eg. function q on line 23 of the last part.
Change function name findDistances to getDistanceMatrix.
Add more comments.

Errors:
Adding the function create_phylogenetic_guide_tree is giving an error "Cannot `convert` an object of type Main.TreeMod.Tree to
an object of type Main.TreeMod.Tree" which is stupid because both the objects are of the same type (to be inferred as Julia is awesome!!!!).
Trace:

MethodError: Cannot `convert` an object of type Main.TreeMod.Tree to an object of type Main.TreeMod.Tree
Closest candidates are:
  convert(::Type{T}, !Matched::T) where T at essentials.jl:154
  Main.TreeMod.Tree(::Any, !Matched::Any) at In[139]:11

Stacktrace:
 [1] setindex!(::Dict{Main.TreeMod.Tree,Float64}, ::Int64, ::Main.TreeMod.Tree) at ./dict.jl:373
 [2] A_Star(::Main.TreeMod.Tree, ::Main.TreeMod.Tree) at ./In[141]:237
 [3] create_phyogenetic_tree(::Array{Float64,2}, ::Bool) at ./In[141]:354
 [4] top-level scope at In[141]:415