Added .jl file which is essentially the entire code with profiling included.
The code expects an input file parameter which has a list of newline seperated sequences.

The time and memory statistics are printed as "name_of_inp_file"_stats.txt.
Elapsed time is the difference between wall clock time before execution of main function and after it.
The MemAlloc field is the total memory allocated in bytes.

Also a nwk file is created as an output.